Make naming scheme for movies work better with plex
previously movie episode numbers were not stored in the name and were stored in a special section at
the end of the name (inside square brackets) only for a human reader. However, for movies stored as multi-part movies this breaks the stacking behavior (since plex requires the names to be otherwise identical except for the part number). In addition some anidb movies have multiple episodes with titles and episode numbers. So storing the episode number explicitly, will guarantee correct ordering and have plex correctly identify the episode number.

In the new format any movie titled 'Complete Movie' or 'Part X of Y' will not store the episode number. Specials or movies with titles will store an episode number.