Introduce Redis as a data store
previously no persistent data and was stored and we were always
re-requesting data from Anidb. While this worked fine, it meant we
were looking up the same data multiple times. We can also do other
useful things by having the data persisted.

This diff introduces Redis as a persistent store, creates a script to
import an xml mylist import to seed the database and then anytime
a file is scanned, this results in updates to the data store.