Refactor script into single purpose processors
To make this script also function as a webapp, a couple of
single purpose processors are created. The file processor can process
a single file end to end. The batch processor takes a batch of files
and now handles the fact that one or more files may be duplicates of
existing files.

A record of every scan is kept that includes which updates were discovered.

Actions at the end of the script are moved into a post_batch actions
section