Data model Improvements: Clear Keys, Store JSON
Earlier when an existing object was saved, if an attribute was nil,
the data in redis was not being cleared. Now with this change the value
is explicitly removed from the redis hash.

Also now with hash_attrs, there is support to store a hash as json.
BatchData was storing a json and is refactored to use this capability.